+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## [2.0.1] - 07-30-23
+
+- Added default-on option to sync video player to run build hooks
+- Moved sync video player sync options under an "Advanced" sub-section.
+- Removed reference to Video Manager and Audio Manager from Sync Video Player
+- Video Manager and Audio Manager objects register themselves with a video player on startup
+- Updated sync video player integrity check to look for none or duplicate managers referencing it
+
 ## [2.0.0] - 07-29-23
 
 - First post-VPM release